refactor: made helper for session tests
